Change protobuf lens definitions
Whenever you have a `Lens' a b` (Or something similar), changing the b inside
the a won’t modify the structure of a.  That is, when you change the content of
a maybe value, you cannot change whether it is Just or Nothing:

Just 1  & .~ 2 --> Just 2
Nothing & .~ 2 --> Nothing

The same applies to the lenses we exposed, e.g. `open_channel`.